Type_Prefix changed to Prefix_Prefix, bug fixes, segments for memory values

+++ Architecture Compiler/compiler/Prefix.cpp
@@ -1,7 +1,7 @@
 #include "Prefix.hpp"
 #include "BytePackage.hpp"
 
-Prefix::Prefix(uint16_t prefix) :
+Prefix::Prefix(uint32_t prefix) :
 	m_Prefix(prefix)
 {
 	if (prefix & Type_RexW)
